Rewired and refactored storyboards. Deleted scroll view, and moved everything into CreateDeckVC. Moved and refactored code into CreateDeckVC. Need to fix sections and cells inside Tableview for adding card and viewing a card.
