+++ mNeme/mNeme/View Controllers/CreateCardTableViewCell.swift
@@ -1,0 +1,15 @@
+//
+//  CreateCardTableViewCell.swift
+//  mNeme
+//
+//  Created by Niranjan Kumar on 3/13/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import UIKit
+
+class CreateCardTableViewCell: UITableViewCell {
+
+
+
+}
